Basic function wrapping for Haskell target language.
The calling conventions are quite unhaskelly - both input and output
parameters come in the form of an STRef passed to the function; with there
being no return value in the ST monad (the return type of all functions is
ST s ()).

The handling of return parameters closely mimics the C implementation
(that's why I did it that way).

Input parameters are passed as STRefs because all of the operators at the
moment expect variables to be STRefs rather than raw values.

Whilst this gives lots of flexibility for the called function to mess
around with return values (and even screw with input values at their
source location!), its probably better to end up with a more Haskelly style
in the long run - eg for programmer familiarity, and to take advantage
of Haskell's type checking more fully.