A command-line utility to pipe stdin through the sha256 algorithm.
Following tradition, this does not work.

It appears to output a hash that depends on the length of the input, but
not the actual bits of the input.

To build, first run ./test.sh Haskell (to generate sha256.hs) and then in the
test/cli/ directory, type  "make"

The wrapper code, sha256cli.hs, gives an example of how to use the
present function bindings from normal haskell code.